+++ packages/hydra-indexer-gateway/src/modules/substrate-event/substrate-event.model.ts
@@ -51,6 +51,15 @@ export class SubstrateEvent extends BaseModel {
   @JSONField()
   phase!: AnyJson
 
+  @JSONField()
+  data!: AnyJson
+
+  @JSONField()
+  extrinsicArgs!: AnyJson
+
+  @StringField({ nullable: true })
+  extrinsicHash?: string
+
   @IntField()
   blockNumber!: number